backend/requirements.txt
# --- Core Flask dependencies ---
Flask>=2.1
flask-cors>=3.0
flask-pymongo>=2.3
gunicorn>=20.1.0
python-dotenv>=1.0

# --- Security & Auth ---
passlib[bcrypt]>=1.7
PyJWT>=2.6
Flask-JWT-Extended>=4.4.4

# --- Database & Background Jobs ---
pymongo>=4.0
apscheduler>=3.9

# --- Utilities ---
requests>=2.28
pydantic[email]>=1.10
email-validator>=2.0.0   # <- REQUIRED so Pydantic’s EmailStr works

# --- AI & Forecasting Modules ---
sentence-transformers>=2.2.2
prophet>=1.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# --- Web Scraping ---
beautifulsoup4>=4.12.0
lxml>=4.9.0
